+++ src/main/java/nextstep/auth/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.exception;
+package nextstep.auth;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
