+++ src/test/java/nextstep/auth/application/GithubClientTest.java
@@ -1,6 +1,6 @@
 package nextstep.auth.application;
 
-import nextstep.exception.AuthenticationException;
+import nextstep.auth.AuthenticationException;
 import nextstep.auth.application.dto.GithubProfileResponse;
 import nextstep.utils.GithubResponse;
 import org.assertj.core.api.Assertions;
